+++ accessibility-checker/src-ts/lib/.gitignore
@@ -1,0 +1,1 @@
+common
