--- accessibility-checker/src-ts/lib/reporters/.gitignore
@@ -1,1 +1,0 @@
-genReport.ts
