+++ common/module/src/api-ext/IAbstractAPI.ts
@@ -14,22 +14,23 @@
     limitations under the License.
 *****************************************************************************/
 
-import { eRuleCategory, eRulePolicy, eRulesetType, eToolkitLevel } from "./IReport"
+/**
+ * Interface for writing files
+ */
+export interface IAbstractAPI {
+    /**
+     * This method should ensure that the directory exists and write a file to the outputDir as specified by config.
+     * @param filePath 
+     * @param data 
+     */
+    writeFileSync(filePath: string, data: string | Buffer)
 
-export interface IRuleset {
-    id: string,
-    name: string,
-    category: eRuleCategory,
-    description: string,
-    type?: eRulesetType,
-    checkpoints: Array<{
-        num: string,
-        // See https://github.com/act-rules/act-tools/blob/main/src/data/sc-urls.json
-        scId?: string,
-        // JCH: add name of checkpoint and summary description
-        name: string,
-        wcagLevel: string,
-        summary: string,
-        rules?: Array<{ id: string, level: eRulePolicy, toolkitLevel: eToolkitLevel }>
-    }>
+    loadBaseline(label: string);
+
+    info(...args: any[])
+
+    log(...args: any[])
+
+    error(...args: any[])
+
 }
